+++ components/console/src/components/form/index.ts
@@ -1,0 +1,6 @@
+// Form components exports
+export { SearchAccountByAliasField } from './search-account-by-alias-field'
+export type { SearchAccountByAliasFieldProps, AccountDto } from './search-account-by-alias-field'
+
+// Re-export for backward compatibility
+export { SearchAccountByAliasField as SearchAccountByAlias } from './search-account-by-alias-field'
