+++ components/console/src/components/form/index.tsx
@@ -6,3 +6,4 @@ export * from './state-field'
 export * from './country-field'
 export * from './currency-field'
 export * from './copyable-input-field'
+export * from './search-account-by-alias-field'
